requirements.txt
kafka-python
prometheus-client